refactor(settings): move branding accessors, consts, validators to settings_branding.go (PRD #1021 M2)
